README: Clarify installation instructions
With default Emacs configuration, latest stable (tagged) version of
gptel is available for installation from NonGNU ELPA, so it is not the
latest commit that will be installed if the user has not modified
`package-archives'.

README.org: Make it more clear how to install stable or latest commit,
and mention `use-package'.